back/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
build==1.2.1
cachetools==5.3.3
CD==1.1.0
certifi==2022.9.24
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.3
django-allauth==0.62.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.13.3
google==3.0.0
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.7
numpy==1.24.2
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.0
pgsql==1.2
pillow==10.3.0
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.6
pumpy==1.1.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pygame==2.3.0
pygame-menu==4.4.2
PyJWT==1.7.1
pyparsing==3.1.2
pyperclip==1.8.2
pyproject_hooks==1.0.0
pyserial==3.5
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.1
